move relabeling functionality to its own package
also remove the returned error as it was always nil
